+++ client/src/App.jsx
@@ -4,6 +4,7 @@ import Login from './pages/Login';
 import Home from './pages/Home';
 import AlumniInvite from './pages/AlumniInvite';
 import CompleteProfile from './pages/CompleteProfile';
+import CommunityPosts from './pages/CommunityPosts';
 
 function App() {
   return (
@@ -12,6 +13,7 @@ function App() {
         <Route path="/login" element={<Login />} />
         <Route path="/alumni-invite/:token" element={<AlumniInvite />} />
         <Route path="/complete-profile" element={<CompleteProfile />} />
+        <Route path="/community/:id" element={<CommunityPosts />} />
         <Route path="/home" element={<Home />} />
         <Route path="/" element={<Navigate to="/home" replace />} />
       </Routes>